refactor(behaviors): move `Projectile` implementations together
